+++ structure/public/index.php
@@ -8,6 +8,6 @@
 
 ini_set('display_errors', '1');
 
-$app = require_once(__DIR__ . '../vendor/devvoh/parable/src/Framework/Bootstrap.php');
+$app = require_once(__DIR__ . '/../vendor/devvoh/parable/src/Framework/Bootstrap.php');
 
 $app->run();